+++ src/components/modal-overlay/modal-overlay.jsx
@@ -1,17 +1,34 @@
 import styles from './modal-overlay.module.css';
 import PropTypes from 'prop-types';
+import { useEffect } from 'react';
 
 
-function ModalOverlay({ children, isOpen, onClose }) {
+function ModalOverlay({ isOpen, onClose }) {
+    useEffect(() => {
+        const handleCloseEscape = (evt) => {
+            if (evt.key === 'Escape') {
+                onClose();
+            }
+        };
+        isOpen && document.addEventListener('keydown', handleCloseEscape);
+
+        return () => {
+            document.removeEventListener('keydown', handleCloseEscape);
+        }
+    }, [isOpen, onClose]);
+
+    const handleCloseOverlayClick = (evt) => {
+        evt.stopPropagation();
+        onClose();
+    };
+
     return (
-        <div className={isOpen ? styles.overlay_opened : styles.overlay_closed} onClick={onClose}>
-            {children}
-        </div>
+        isOpen &&
+        <div className={styles.overlay} onClick={handleCloseOverlayClick}></div>
     )
 }
 
 ModalOverlay.propTypes = {
-    children: PropTypes.node.isRequired,
     isOpen: PropTypes.bool.isRequired,
     onClose: PropTypes.func.isRequired
 }
