+++ src/components/modal/modal.module.css
@@ -5,6 +5,11 @@
     box-sizing: border-box;
     border-radius: 40px;
     background-color: rgb(28, 28, 33);
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 2;
 }
 
 .wrapper {
